Setup Revamped, Middle of Add Pane Revamp
Reformatted Setup Function to only create the Tkinter Root, and set up the main menu for the Alexandria program. Each Page should start by initializing the master frame.

Currently working on the BuildAddPane function, which allows users to add more book data to the database. Implementing the LabelsAndEntries module, I streamlined the GUI setup and information extraction. Additionally, added functionality where the user can change the number of input authors.

Currently working on the tools to take the information from the user and input it into the database.